% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmantismod-package.R
\docType{package}
\name{rmantismod-package}
\alias{rmantismod}
\alias{rmantismod-package}
\title{rmantismod: modifieds mantis (GPU)}
\description{
This is a modified version of rmantis.
It has GPU support for 1 CUDA device.
}
\section{Main functions}{

The core functions of \code{mantismod} are:
\itemize{
\item \code{\link{mantis_cudacheck}} - Checks CUDA availability and lists GPU devices
\item \code{\link{mantis_download_weights}} - Downloads the weights
\item \code{\link{mantis_forecast}} - Performs a forecast
}
}

\section{Getting started}{

A typical workflow looks like:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{library(rmantismod)
# Verify your GPU environment first
mantis_cudacheck("/path/to/python")
res <- mantis_forecast
}\if{html}{\out{</div>}}
}

\references{
Any papers, books, or URLs your package is based on.
}
\author{
Wim R.M. Cardoen \email{wcardoen@gmail.com}
}
\keyword{internal}
